LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/database-odd-names-expected.txt

CONSOLE MESSAGE: line 27: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedErrorCallback')
Test IndexedDB odd value datatypes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

opening a database named null
indexedDB.open(testData[nextToOpen].name)
FAIL indexedDB.open(testData[nextToOpen].name) threw exception TypeError: null is not an object (evaluating 'indexedDB.open')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
